==> notebooks/11_backfill_feature_store.ipynb <==
%reload_ext autoreload
%autoreload 2

# # Fix for Hopsworks Kafka Producer issue
# import importlib
# import sys

# # Make sure kafka-python is installed
# try:
#     from kafka import KafkaProducer
# except ImportError:
#     raise ImportError("Please install kafka-python: pip install kafka-python")

# # Monkey patch the missing Producer
# import hsfs.core.kafka_producer
# if not hasattr(hsfs.core.kafka_producer, "Producer"):
#     hsfs.core.kafka_producer.Producer = KafkaProducer

HOPSWORKS_PROJECT_NAME = 'msd_taxi_demand'

import os
import sys

# Add the parent directory (project root) to the Python path
project_root = os.path.abspath(os.path.join(os.getcwd(), '..'))
if project_root not in sys.path:
    sys.path.insert(0, project_root)
    
# Now try the import
from dotenv import load_dotenv
from src.paths import PARENT_DIR

# load key-value pairs from .env file
load_dotenv(PARENT_DIR / '.env')
HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

from_year = 2022
to_year = datetime.now().year
print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    
    # download data for the whole year
    rides_one_year = load_raw_data(year)
    
    # append rows
    rides = pd.concat([rides, rides_one_year])

print(f'{len(rides)=:,}')

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

# string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype(int) // 10**6

import hopsworks

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_ts'],
    event_time='pickup_ts',
)

import hsfs
print(hsfs.__version__)

try:
    from confluent_kafka import Producer
    print("Producer class is available")
except ImportError:
    print("Producer class is not available")

feature_group.insert(ts_data, write_options={"wait_for_job": False})
